package pending
Réorganisation du code
Renommé DoubleHeaderAdapter en WorkAdapter (plus cohérent)
