+++ app/src/main/AndroidManifest.xml
@@ -33,6 +33,13 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".pages.MainPage" />
         </activity>
+        <activity
+            android:name=".pages.CommentPage"
+            android:label="@string/title_commentpage">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".pages.MainPage" />
+        </activity>
     </application>
 
 </manifest>
