+++ app/src/main/java/fr/agendapp/app/objects/FusionList.java
@@ -9,7 +9,7 @@ import android.widget.TextView;
 import java.util.Calendar;
 
 import fr.agendapp.app.R;
-import fr.agendapp.app.factories.PendMERGE;
+import fr.agendapp.app.pending.PendMERGE;
 
 public class FusionList {
