+++ app/src/main/java/fr/agendapp/app/pending/PendMERGE.java
@@ -1,4 +1,5 @@
-package fr.agendapp.app.factories;
+package fr.agendapp.app.pending;
+
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -8,40 +9,30 @@ import java.util.List;
 import java.util.ListIterator;
 
 import fr.agendapp.app.App;
-import fr.agendapp.app.objects.Comment;
+import fr.agendapp.app.factories.ParseFactory;
 
 /**
  * @author Dylan Habans
  */
-public class PendCOMM extends Pending {
+public class PendMERGE extends Pending {
 
-    private static List<PendCOMM> pending;
-    private static String name = "pendCOMM";
+    private static List<PendMERGE> pending;
+    private static String name = "pendMERGE";
 
-    private int id;
-    private String comment;
+    private int[] ids;
 
-    /**
-     * @param id ID
-     * @param comment Commentaire
-     * Constructeur de PendCOMM
-     */
-    private PendCOMM(Context context, int id, String comment) {
-        this.id = id;
-        this.comment = comment;
+    public PendMERGE(Context context, int[] ids) {
+        this.ids = ids;
         pending.add(this);
-        PendCOMM.saveList(context);
+        PendMERGE.saveList(context);
     }
 
-    public PendCOMM(Context context, Comment c) {
-        this(context, c.getId(), c.getText());
-    }
 
     /**
-     * @return représentation JSON de la liste d'actions PendCOMM
+     * @return représentation JSON de la liste d'actions PendDELc
      */
     static String getList() {
-        ListIterator<PendCOMM> i = pending.listIterator();
+        ListIterator<PendMERGE> i = pending.listIterator();
         String json = "[";
         while (i.hasNext()) {
             json += i.next();
@@ -53,7 +44,7 @@ public class PendCOMM extends Pending {
 
     static void initList(Context context) {
         SharedPreferences preferences = context.getSharedPreferences(App.TAG, Context.MODE_PRIVATE);
-        pending = ParseFactory.parsePendCOMM(preferences.getString(name, "[]"));
+        pending = ParseFactory.parsePendMERGE(preferences.getString(name, "[]"));
     }
 
     static void saveList(Context context) {
@@ -76,17 +67,14 @@ public class PendCOMM extends Pending {
         return name;
     }
 
-    /**
-     * @return représentation JSON de l'action PendCOMM
-     */
     public String toString() {
-
-        String json = "{";
-        json += "\"id\":" + id + ",";
-        json += "\"content\": {" +
-                "\"texte\": \"" + comment + "\"";
-        json += "}";
-
+        String json = "[";
+        for (int i = 0; i < ids.length; i++) {
+            json += ids[i];
+            json += ",";
+        }
+        json = json.substring(0, json.length() - 1); // supprime la dernière virgule
+        json += "]";
         return json;
     }
 }
